+++ example/lib/data/mock_universal_ble.dart
@@ -5,7 +5,7 @@ import 'package:universal_ble/universal_ble.dart';
 
 /// Mock implementation of [UniversalBlePlatform] for testing
 class MockUniversalBle extends UniversalBlePlatform {
-  final _mockBleScanResult = BleScanResult(
+  final _mockBleDevice = BleDevice(
     name: 'MockDevice',
     deviceId: 'MockDeviceId',
     rssi: 50,
@@ -23,11 +23,8 @@ class MockUniversalBle extends UniversalBlePlatform {
   ]);
 
   @override
-  Future<void> startScan({
-    ScanFilter? scanFilter,
-  }) async {
-    onScanResult?.call(_mockBleScanResult);
-  }
+  Future<void> startScan({ScanFilter? scanFilter}) async =>
+      onScanResult?.call(_mockBleDevice);
 
   @override
   Future<void> stopScan() async {}
@@ -59,8 +56,7 @@ class MockUniversalBle extends UniversalBlePlatform {
   }
 
   @override
-  Future<List<BleScanResult>> getConnectedDevices(
-      List<String>? withServices) async {
+  Future<List<BleDevice>> getSystemDevices(List<String>? withServices) async {
     return [];
   }
 
@@ -106,5 +102,10 @@ class MockUniversalBle extends UniversalBlePlatform {
   @override
   Future<void> unPair(String deviceId) async {
     onPairingStateChange?.call(deviceId, false, null);
+  }
+
+  @override
+  Future<bool> isConnected(String deviceId) {
+    throw UnimplementedError();
   }
 }