+++ lib/src/models/ble_property.dart
@@ -1,22 +1,18 @@
 enum BleInputProperty {
-  disabled(0),
-  notification(1),
-  indication(2);
+  disabled,
+  notification,
+  indication;
 
-  final int value;
-  const BleInputProperty(this.value);
+  const BleInputProperty();
 
-  factory BleInputProperty.parse(int value) =>
-      BleInputProperty.values.firstWhere((element) => element.value == value);
+  factory BleInputProperty.parse(int index) => BleInputProperty.values[index];
 }
 
 enum BleOutputProperty {
-  withResponse(0),
-  withoutResponse(1);
+  withResponse,
+  withoutResponse;
 
-  final int value;
-  const BleOutputProperty(this.value);
+  const BleOutputProperty();
 
-  factory BleOutputProperty.parse(int value) =>
-      BleOutputProperty.values.firstWhere((element) => element.value == value);
+  factory BleOutputProperty.parse(int index) => BleOutputProperty.values[index];
 }