+++ pigeon/universal_ble.dart
@@ -73,9 +73,11 @@ abstract class UniversalBlePlatformChannel {
   void unPair(String deviceId);
 
   @async
-  List<UniversalBleScanResult> getConnectedDevices(
+  List<UniversalBleScanResult> getSystemDevices(
     List<String> withServices,
   );
+
+  bool isConnected(String deviceId);
 }
 
 /// Native -> Flutter